requirements.txt
streamlit
pillow
pillow-heif
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib